refactor: move question field components into subfolder
