+++ src/client/components/builder/index.ts
@@ -1,10 +1,6 @@
 export { createBuilderField } from './BuilderField'
 export { Navbar } from './Navbar'
 export { FloatingToolbar } from './FloatingToolbar'
-export { NumericField } from './NumericField'
-export { RadioField } from './RadioField'
-export { CheckboxField } from './CheckboxField'
-export { TitleField } from './TitleField'
 export { QuestionsTab } from './QuestionsTab'
 export { LogicTab } from './LogicTab'
 export { PreviewTab } from './PreviewTab'
