+++ src/client/components/builder/questions/CheckboxField.tsx
@@ -12,11 +12,10 @@ import {
   CheckboxGroup,
 } from '@chakra-ui/react'
 
-import * as checker from '../../../types/checker'
-import { createBuilderField, QuestionFieldComponent } from './BuilderField'
-import { useCheckerContext } from '../../contexts'
-
-import { BuilderActionEnum, ConfigArrayEnum } from '../../../util/enums'
+import * as checker from '../../../../types/checker'
+import { createBuilderField, QuestionFieldComponent } from '../BuilderField'
+import { useCheckerContext } from '../../../contexts'
+import { BuilderActionEnum, ConfigArrayEnum } from '../../../../util/enums'
 
 const InputComponent: QuestionFieldComponent = ({ field, index }) => {
   const { description } = field
