+++ src/client/components/builder/questions/RadioField.tsx
@@ -12,10 +12,10 @@ import {
   RadioGroup,
 } from '@chakra-ui/react'
 
-import * as checker from '../../../types/checker'
-import { useCheckerContext } from '../../contexts'
-import { createBuilderField, QuestionFieldComponent } from './BuilderField'
-import { BuilderActionEnum, ConfigArrayEnum } from '../../../util/enums'
+import * as checker from '../../../../types/checker'
+import { useCheckerContext } from '../../../contexts'
+import { createBuilderField, QuestionFieldComponent } from '../BuilderField'
+import { BuilderActionEnum, ConfigArrayEnum } from '../../../../util/enums'
 
 const InputComponent: QuestionFieldComponent = ({ field, index }) => {
   const { description } = field
