+++ src/client/components/builder/questions/index.ts
@@ -1,0 +1,4 @@
+export { TitleField } from './TitleField'
+export { NumericField } from './NumericField'
+export { RadioField } from './RadioField'
+export { CheckboxField } from './CheckboxField'
